backend/requirements.txt
# backend\requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-multipart==0.0.9
pymupdf==1.24.3
python-docx==1.1.2
openai==1.30.3
google-genai>=1.0.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
pymysql==1.1.0
cryptography==42.0.7
pymysql
